backend/requirements.txt
google-cloud-core>=1.0.0
google-cloud-storage>=1.13.2
google-cloud-logging
grpcio
google-cloud-firestore
gunicorn==19.9.0
flask==0.12.4
pyjwt==1.7.1
flask-cors==3.0.7
firebase-admin
google-auth==1.6.2
requests==2.21.0
requests-toolbelt==0.9.1
Werkzeug<0.13.0,>=0.12.0
